Refactored the calibration script
Modularized Arduino code, into communications, controller and
calibration.

Defined the message struct (see comms.h) to have 32 bytes,
which is the maximum supported by the Wire library.

Next improvement will be a timeout for when the calibration is stuck.

Controller consists of a C++ class (to be developed), for now it just has the method compute_lux() to easily convert sensor value to LUX.