Arduino code ready for shipping
Consensus works well now.
Added a feature that resets the integral term if a new dimming is
desired.
Still, there is an issue, if there is a heavy disturbance, the feedback
controller will compensate too much on actuation, such that the optimal
dimmings won't be restored when the disturbance goes away.